Fix of the unit tests after MySQL grammar update + some TUT improvements
MySQL grammar + parser:
- Merged all latest changes of the server grammar from mysql-trunk into our ANTLR grammar.
- Removed obsolete ANTLRv3 parser grammars.
- Updated the parser module + object listeners.
- Added ability to add more than one lexer token per step, which allows to emit multiple tokens while recognizing a single construct (e.g. ".id" -> 2 tokens: dot and ID).
- Changed number + identifier lexer rules to recognize special constructs like ".1u" (which must be handled as identifier with a leading dot, not a float and a "u" appendix). Added new unit tests for that.
- Corrected binary and hex number lexing (they only accept lower case b and x prefixes). Added new unit tests for that as well.
- Removed old parser from TUT helpers and use now the new parser module for splitting + running queries. Requires to change some shutdown code as d-tors of test classes are run during static de-init, where all our modules are long gone.
- Updated parser generation scripts to use the latest ANTLR 4.6 version.

Code completion:
- Added new code completion core class to parser lib.
- Added code completion function to parser module.
- Re-enabled code completion in SQL editors and updated it to use the new parser module function.

Other:
- Chnaged a few tests that used invalid MySQL code and returned different results than expected.
- Raised default min server version in SQL editor form to 5.7.
- Some adjustments in the context help class due to grammar changes.
- Updated grtdb data type tests to take the new (date)time precisions into account.
- mforms: bug fix in code editor config (for tests we use a scintilla mockup which doesn't really send commands)
- Update sql code generation so that "drop view if exists v1;" appears before "drop table if exists v1;" (for view placehoders) because MySQL server 8+ issue a server error (instead of a warning as before) if we try to drop a view is actually a table.
- The XCode scheme for the new mtemplate engine is now shared.
- Updated the WB XCode project with the new files.

+++ backend/wbpublic/sqlide/unit-tests/sql_editor.cpp
@@ -57,9 +57,9 @@ TEST_FUNCTION(1)
   parsers::MySQLParserServices::Ref services = parsers::MySQLParserServices::get();
   parsers::MySQLParserContext::Ref parser = services->createParserContext(rdbms->characterSets(), version, "", 1);
 	ensure("failed to retrieve RDBMS list", rdbms_list.is_valid());
-	for (int n= 0, count= rdbms_list.count(); n < count; ++n)
+	for (size_t n = 0, count = rdbms_list.count(); n < count; ++n)
 	{
-		db_mgmt_RdbmsRef rdbms= rdbms_list[n];
+		db_mgmt_RdbmsRef rdbms = rdbms_list[n];
 		MySQLEditor::Ref sql_editor = MySQLEditor::create(parser, parser);
 		ensure(("failed to get sql editor for " + rdbms->name().toString() + " RDBMS").c_str(), (NULL != sql_editor.get()));
 	}
